Move test_remote_submission_worker.py to proper location
