+++ tests/unit/remoteworker/test_remote_worker.py
@@ -1,11 +1,16 @@
 import mock
+import os
+import responses
+import shutil
 
 from unittest import TestCase
 
 from scripts.workers.remote_submission_worker import (
+    create_dir,
     make_request,
     get_message_from_sqs_queue,
     delete_message_from_sqs_queue,
+    download_and_extract_file,
     get_submission_by_pk,
     get_challenge_phases_by_challenge_pk,
     get_challenge_by_queue_name,
@@ -141,3 +146,45 @@ class URLFormatTestCase(BaseTestClass):
         expected_url = "http://testserver:80{}".format(url)
         returned_url = return_url_per_environment(url)
         self.assertEqual(returned_url, expected_url)
+
+
+class TestDownloadAndExtractFile(BaseTestClass):
+    def setUp(self):
+        super(TestDownloadAndExtractFile, self).setUp()
+
+        self.download_dir = "/tmp/evalai/"
+        self.download_filename = "file"
+        self.download_path = "{}{}".format(self.download_dir, self.download_filename)
+
+        self.url = "http://example.com"
+        self.body = """{"content": "test-data"}"""
+        responses.add(
+            responses.GET,
+            self.url,
+            body=self.body,
+            status=200
+        )
+
+        create_dir(self.download_dir)
+
+    def tearDown(self):
+        shutil.rmtree(self.download_dir)
+
+    @responses.activate
+    def test_download_and_extract_file_with_correct_url(self):
+        assert os.path.exists(self.download_dir)
+
+        download_and_extract_file(self.url, self.download_path)
+        with open(self.download_path, "r") as f:
+            assert f.read() == self.body
+
+    @mock.patch("scripts.workers.remote_submission_worker.logger.error")
+    def test_download_and_extract_file_with_incorrect_url(self, mock_logger):
+        assert os.path.exists(self.download_dir)
+
+        self.url = "invalid-url"
+        download_and_extract_file(self.url, "")
+        mock_logger.assert_called_with(
+            "Failed to fetch file from {}, error "
+            "Invalid URL 'invalid-url': No schema supplied. Perhaps you meant http://invalid-url?".format(self.url)
+        )
